#136 moving the search job into the utility package
