+++ applications/channel/views/.classpath
@@ -8,4 +8,5 @@
 	<classpathentry combineaccessrules="false" kind="src" path="/core-framework"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/core-ui"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/applications-channelfinder"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/applications-channel-utility"/>
 </classpath>
